+++ go/proto/gen.go
@@ -15,11 +15,11 @@ import (
 
 // Any proto which will be marshalled as a single capnp message needs to be listed here.
 var RootTypes = []string{
+	"CtrlPld",
 	"PathSegment",
 	"RevInfo",
-	"CtrlPld",
-	"CtrlPldOuter",
 	"SCIONDMsg",
+	"SignedCtrlPld",
 }
 
 func main() {
